+++ packages/demo/packages/frontend/src/api/client.ts
@@ -1,5 +1,4 @@
-
-import * as RealmWeb from 'realm-web'
+import * as Flowerbase from '@flowerforce/flowerbase-client'
 
 const DB_NAME = import.meta.env.VITE_DB_NAME as string
 const APP_ID = import.meta.env.VITE_APP_ID as string
@@ -10,7 +9,7 @@ const appConfig = {
   baseUrl: import.meta.env.VITE_SERVER_URL,
 }
 
-export const app = new RealmWeb.App(appConfig)
+export const app = new Flowerbase.App(appConfig)
 export const db = (dbname: string = DB_NAME) =>
   app.currentUser!.mongoClient('mongodb-atlas').db(dbname)
-export const Realm = RealmWeb
+export const Realm = Flowerbase